Action/Server: Take the main search result link from reversed module
Previously only the module list was reversed, but the main link to a
search result would still point to the first module in the
non-reversed list.

This uses the logic for both the `showFroms` and the main result loop.

Note that the order of *packages* is still pretty much arbitrary, so
the user might still land in a package other than where the symbol was
originally defined.

But for example in the case of `find`, the `base` package is most
often displayed first, and the `find` from `Data.Foldable` will now
take precedence of the reexport from the reexport in `Data.List`.